Reorganize packet modules
