+++ crates/valence_protocol/src/lib.rs
@@ -102,7 +102,6 @@ mod block_pos;
 mod bounded;
 mod byte_angle;
 mod codec;
-pub mod command;
 pub mod enchant;
 pub mod entity_meta;
 pub mod ident;
@@ -110,10 +109,7 @@ mod impls;
 mod inventory;
 mod item;
 pub mod packets;
-pub mod particle;
-pub mod player_list;
 mod raw_bytes;
-pub mod recipe;
 pub mod text;
 pub mod translation_key;
 pub mod types;
