+++ crates/valence_protocol/src/packets/s2c/declare_recipes.rs
